MOD: organize view folders
